Resolved Final Link Issue and Metadata is Now Pulled from Existing Yaml Files
The motivation behind the final link change is self-evident. It will allow the user to go directly to the uploaded file without downloading it. The placement of the metadata in the existing yaml file makes the repository less complicated with unnecessary files and makes the information easier to find overall.